reader tests \u escaping utf8 support

+++ c/src/tests.h
@@ -9,7 +9,7 @@
 #define malloc test_malloc
 #define free test_free
 
-void *test_malloc(size_t size);
+void *test_malloc(int size);
 void test_free(void *);
 void fail(const char *s);
